requirements.txt
aniso8601==8.1.0
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
crayons==0.3.1
Django==3.1.5
falcon==2.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
hug==2.6.1
idna==2.10
image==1.5.33
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
messagebird==1.5.0
numpy==1.19.4
piexif==1.1.3
Pillow==6.2.2
pycnic==0.1.4
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
stegano==0.9.8
toml==0.10.2
twilio==6.48.0
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
